requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.100.0
    # via tiny-boss (pyproject.toml)
anyio==4.13.0
    # via
    #   anthropic
    #   httpx
    #   openai
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
cryptography==48.0.0
    # via google-auth
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.18.0
    # via anthropic
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.30.3
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.196.0
    # via google-generativeai
google-auth==2.52.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.4.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via tiny-boss (pyproject.toml)
googleapis-common-protos==1.75.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.80.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   anthropic
    #   openai
idna==3.13
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.14.0
    # via
    #   anthropic
    #   openai
openai==2.36.0
    # via tiny-boss (pyproject.toml)
proto-plus==1.28.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.13.4
    # via
    #   anthropic
    #   google-generativeai
    #   openai
pydantic-core==2.46.4
    # via pydantic
pyparsing==3.3.2
    # via httplib2
requests==2.33.1
    # via google-api-core
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
tqdm==4.67.3
    # via
    #   google-generativeai
    #   openai
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   google-generativeai
    #   grpcio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.7.0
    # via requests
